pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lexia-cr"
version = "0.1.0"
description = "LexIA — Judicial Response Letter automation (Carta-Resposta) without UiPath"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27,<1.0",
    "pydantic>=2.9,<3.0",
    "pydantic-settings>=2.6,<3.0",
    "python-dotenv>=1.0",
    "structlog>=24.0,<26.0",
    "rich>=13.9",
    "typer>=0.12,<1.0",
    "databricks-sql-connector>=3.0,<4.0",
    "openai>=1.50,<3.0",
    "gspread>=6.0,<7.0",
    "google-auth>=2.29,<3.0",
    "slack-sdk>=3.27,<4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "ruff>=0.7",
]

[project.scripts]
lexia = "lexia.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]
